backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
babel==2.16.0
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
distro==1.9.0
docxcompose==1.4.0
docxtpl==0.19.0
Flask==3.0.3
Flask-Cors==5.0.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.0
lxml==5.3.0
MarkupSafe==3.0.0
openai==1.57.2
packaging==24.1
pydantic==2.9.2
pydantic_core==2.23.4
python-docx==1.1.2
python-dotenv==1.0.1
redis==5.2.1
rq==2.0.0
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
Werkzeug==3.0.4
# TIA Generator Backend Requirements
# Core dependencies
fastapi>=0.101.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.24.1

# OpenAI API
openai>=1.2.0
tenacity>=8.2.2

# Document generation
python-docx>=0.8.11
docxtpl>=0.16.7

# Caching
redis>=4.6.0
orjson>=3.9.0

# Performance and async
asyncio>=3.4.3
aiofiles>=23.1.0

# Middleware
starlette>=0.27.0

# Production deployment
gunicorn>=21.2.0
